Extraccion de SlowAssPlayer
esta en un archivo aparte
