+++ nbproject/private/private.xml
@@ -2,6 +2,8 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group/>
+        <group>
+            <file>file:/C:/Users/dell/Documents/GitHub/Maze/MazeImport/src/mazeimport/GameGui.java</file>
+        </group>
     </open-files>
 </project-private>
